+++ src/main/java/frc/robot/commands/autonomous/BalanceOnChargeStation.java
@@ -2,27 +2,42 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.drivetrain;
+package frc.robot.commands.autonomous;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.DriveTrain;
 
 public class BalanceOnChargeStation extends CommandBase {
+  
+  DriveTrain driveTrain;
+
   /** Creates a new BalanceOnChargeStation. */
-  public BalanceOnChargeStation() {
+  public BalanceOnChargeStation(DriveTrain driveTrain) {
     // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(driveTrain);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+    driveTrain.dtInit();
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    // balancing logic:
+    // if pitch is lower than threshold amount:
+    // drive forward
+    // elif pitch is greater than threshold amount:
+    // drive backwards
+  }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    driveTrain.stopAllMotors();
+  }
 
   // Returns true when the command should end.
   @Override
